Fix for window not always closing
There is a bug where, for some people, self.Close in the onCancel function will not work. This will effectively make it so the window cannot be closed at all. Not with the cancel button, the ok button, or the x button on the window. However, like in the dlg_settings file, it uses evt.Skip() to close their windows and it works perfectly. Appending it after the self.Close() function seems to fix this bug. The onCancel function ties into all close functions so when someone hits ok or the x button, not just the cancel button, onCancel is called.